[api][spark][docs] Remove policy replacement APIs
