--- paimon-spark/paimon-spark-common/src/main/java/org/apache/paimon/spark/procedure/CreateOrReplacePolicyBaseProcedure.java
@@ -1,83 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.paimon.spark.procedure;
-
-import org.apache.paimon.management.DataPolicy;
-import org.apache.paimon.management.PolicyType;
-
-import org.apache.spark.sql.catalyst.InternalRow;
-import org.apache.spark.sql.connector.catalog.TableCatalog;
-import org.apache.spark.sql.types.DataTypes;
-import org.apache.spark.sql.types.Metadata;
-import org.apache.spark.sql.types.StructField;
-import org.apache.spark.sql.types.StructType;
-
-import static org.apache.spark.sql.types.DataTypes.StringType;
-
-/** Shared implementation for creating or replacing a table data policy. */
-abstract class CreateOrReplacePolicyBaseProcedure extends BasePolicyProcedure {
-
-    private static final ProcedureParameter[] PARAMETERS =
-            new ProcedureParameter[] {
-                ProcedureParameter.required("database", StringType),
-                ProcedureParameter.required("table", StringType),
-                ProcedureParameter.required("policy_type", StringType),
-                ProcedureParameter.required("principal", StringType),
-                ProcedureParameter.optional("predicate_json", StringType),
-                ProcedureParameter.optional("on_column", StringType),
-                ProcedureParameter.optional("transform_json", StringType)
-            };
-
-    private static final StructType OUTPUT_TYPE =
-            new StructType(
-                    new StructField[] {
-                        new StructField("result", DataTypes.BooleanType, false, Metadata.empty())
-                    });
-
-    protected CreateOrReplacePolicyBaseProcedure(TableCatalog tableCatalog) {
-        super(tableCatalog);
-    }
-
-    @Override
-    public ProcedureParameter[] parameters() {
-        return PARAMETERS;
-    }
-
-    @Override
-    public StructType outputType() {
-        return OUTPUT_TYPE;
-    }
-
-    @Override
-    public InternalRow[] call(InternalRow args) {
-        DataPolicy policy =
-                policy(
-                        args.getString(0),
-                        args.getString(1),
-                        enumValue(args.getString(2), PolicyType.class, PARAMETERS[2].name()),
-                        args.getString(3),
-                        args.isNullAt(4) ? null : args.getString(4),
-                        args.isNullAt(5) ? null : args.getString(5),
-                        args.isNullAt(6) ? null : args.getString(6));
-        writePolicy(policy);
-        return new InternalRow[] {newInternalRow(true)};
-    }
-
-    protected abstract void writePolicy(DataPolicy policy);
-}
