+++ paimon-spark/paimon-spark-common/src/main/java/org/apache/paimon/spark/procedure/CreatePolicyProcedure.java
@@ -19,19 +19,69 @@
 package org.apache.paimon.spark.procedure;
 
 import org.apache.paimon.management.DataPolicy;
+import org.apache.paimon.management.PolicyManagement.PolicyAlreadyExistException;
+import org.apache.paimon.management.PolicyType;
 
+import org.apache.spark.sql.catalyst.InternalRow;
 import org.apache.spark.sql.connector.catalog.TableCatalog;
+import org.apache.spark.sql.types.DataTypes;
+import org.apache.spark.sql.types.Metadata;
+import org.apache.spark.sql.types.StructField;
+import org.apache.spark.sql.types.StructType;
+
+import static org.apache.spark.sql.types.DataTypes.StringType;
 
 /** Creates a table row-filter or column-mask policy. */
-public class CreatePolicyProcedure extends CreateOrReplacePolicyBaseProcedure {
+public class CreatePolicyProcedure extends BasePolicyProcedure {
+
+    private static final ProcedureParameter[] PARAMETERS =
+            new ProcedureParameter[] {
+                ProcedureParameter.required("database", StringType),
+                ProcedureParameter.required("table", StringType),
+                ProcedureParameter.required("policy_type", StringType),
+                ProcedureParameter.required("principal", StringType),
+                ProcedureParameter.optional("predicate_json", StringType),
+                ProcedureParameter.optional("on_column", StringType),
+                ProcedureParameter.optional("transform_json", StringType)
+            };
+
+    private static final StructType OUTPUT_TYPE =
+            new StructType(
+                    new StructField[] {
+                        new StructField("result", DataTypes.BooleanType, false, Metadata.empty())
+                    });
 
     private CreatePolicyProcedure(TableCatalog tableCatalog) {
         super(tableCatalog);
     }
 
     @Override
-    protected void writePolicy(DataPolicy policy) {
-        policyManagement().createPolicy(policy);
+    public ProcedureParameter[] parameters() {
+        return PARAMETERS;
+    }
+
+    @Override
+    public StructType outputType() {
+        return OUTPUT_TYPE;
+    }
+
+    @Override
+    public InternalRow[] call(InternalRow args) {
+        DataPolicy policy =
+                policy(
+                        args.getString(0),
+                        args.getString(1),
+                        enumValue(args.getString(2), PolicyType.class, PARAMETERS[2].name()),
+                        args.getString(3),
+                        args.isNullAt(4) ? null : args.getString(4),
+                        args.isNullAt(5) ? null : args.getString(5),
+                        args.isNullAt(6) ? null : args.getString(6));
+        try {
+            policyManagement().createPolicy(policy);
+        } catch (PolicyAlreadyExistException e) {
+            throw new RuntimeException(e);
+        }
+        return new InternalRow[] {newInternalRow(true)};
     }
 
     public static ProcedureBuilder builder() {
